UnitMultiple accepts all Units
Change UnitMultiple so that the unit does not have to be a SingularUnit:
https://fbresearch.atlassian.net/browse/OM-39

This should make UnitMuliples like 100g possible, where not prefix is used. For instance for Units like g/100g.

This means changing the type of the Unit in UnitMultiple to Unit instead of SingularUnit and changing the documentation.