Add tests for conversion of dimensionless and base units
Compares the result of metre and an other base unit (zaphoid) with the same base dimension. This should result in a conversion exception because the units are unrelated and cannot be converted into each other. This fails at the moment (they are converted) --> [OM-56]

Dimensionless units are also tested. An exception should be thrown when converting between m/m and s/s as this conversion is also not possible. Testing for dimension fails as they are both dimensionless. --> [OM-52]