Added getter for suggested units
Does not yet use Systems of units because they do not yet exist.
The sorted list is sorted according to the 10^e exponent (abs(e)) and on the difference with the 10 base of th source and target units.
For 0.025 m the unit cm will be suggested because cm is in the same base 10, while inch is not.